requirements.txt
Bottleneck @ file:///home/task_176193795131750/conda-bld/bottleneck_1761937999296/work
brotlicffi @ file:///home/task_176496126124857/conda-bld/brotlicffi_1764961326114/work
certifi @ file:///home/task_176765916174011/croot/certifi_1767659183877/work/certifi
cffi @ file:///home/task_176183214835616/conda-bld/cffi_1761832774688/work
charset-normalizer @ file:///home/task_176174481524083/conda-bld/charset-normalizer_1761744828711/work
click @ file:///home/task_176433224228540/conda-bld/click_1764332256338/work
idna @ file:///home/task_176191175488988/conda-bld/idna_1761911981595/work
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159249290440/conda-bld/mkl_fft_1761592897595/work
mkl_random @ file:///home/task_176159294306005/conda-bld/mkl_random_1761592958759/work
numexpr @ file:///home/task_176216555892389/conda-bld/numexpr_1762165570527/work
numpy @ file:///home/task_176397996305634/conda-bld/numpy_and_numpy_base_1763980680623/work/dist/numpy-2.3.5-cp313-cp313-linux_x86_64.whl#sha256=78c581a960868ba31a535f8c54fa0be861dd2d9db1419e7218ace9db5377e772
pandas @ file:///home/task_177064302366852/croot/pandas_1770643078087/work
psycopg2-binary==2.9.11
pycparser @ file:///home/task_175749564936714/conda-bld/pycparser_1757496040682/work
PySocks @ file:///home/task_176175235884818/conda-bld/pysocks_1761753009495/work
python-dateutil @ file:///work/perseverance-python-buildout/croot/python-dateutil_1728385355105/work
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
pytz @ file:///croot/pytz_1752135852232/work
requests @ file:///home/task_176235918788486/conda-bld/requests_1762359591632/work
setuptools==80.9.0
six @ file:///croot/six_1744271502820/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///home/task_176496463239160/conda-bld/urllib3_1764964948143/work
wheel==0.45.1
